Continuing development.

+++ src/index.rs
@@ -1,18 +1,9 @@
-//! # Indexing Module
-//!
-//! Provides the SQLite-based indexing implementation for `LumeEngine`, handling
-//! message metadata retrieval, user authentication, and access control lists (ACLs).
-
 use crate::{security, storage::MailHeader, LumeEngine, LumeError};
 use rusqlite::{params, Connection, Result as SqlResult};
 
 impl LumeEngine {
-    /// Initializes the SQLite database schema.
-    ///
-    /// Creates the `messages` table for metadata and the `users` table for ACL tracking.
     pub fn init_db(db_mutex: &std::sync::Mutex<Connection>) -> SqlResult<()> {
         let db = db_mutex.lock().unwrap();
-        // Mail metadata table
         db.execute(
             "CREATE TABLE IF NOT EXISTS messages (
                 message_id TEXT PRIMARY KEY,
@@ -25,7 +16,6 @@ impl LumeEngine {
             [],
         )?;
 
-        // Secure Users table
         db.execute(
             "CREATE TABLE IF NOT EXISTS users (
                 acl_id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -42,9 +32,6 @@ impl LumeEngine {
         Ok(())
     }
 
-    /// Registers a new user with a securely hashed password.
-    ///
-    /// Returns the newly generated `acl_id` for the user.
     pub fn register_user(
         &self,
         username: &str,
@@ -62,26 +49,20 @@ impl LumeEngine {
         Ok(acl_id)
     }
 
-    /// Authenticates a user and returns their `acl_id` if successful.
-    ///
-    /// To mitigate timing and enumeration attacks, this function performs a dummy
-    /// Argon2id hash if the user is not found in the database.
     pub fn authenticate_user(
         &self,
         username: &str,
         plaintext_password: &str,
     ) -> Result<u64, LumeError> {
-        // Scope the database interaction so the lock and statements
-        // are automatically dropped before the expensive crypto operations.
         let auth_data = {
             let db = self.db.lock().unwrap();
             let mut stmt =
                 db.prepare("SELECT acl_id, password_hash FROM users WHERE username = ?1")?;
             let mut rows = stmt.query(params![username])?;
             if let Some(row) = rows.next()? {
-                let acl_id: u64 = row.get(0)?;
+                let acl_id: i64 = row.get(0)?;
                 let hash: String = row.get(1)?;
-                Some((acl_id, hash))
+                Some((acl_id as u64, hash))
             } else {
                 None
             }
@@ -92,20 +73,12 @@ impl LumeEngine {
                 return Ok(acl_id);
             }
         } else {
-            // Mitigate timing/enumeration attacks by forcing the process
-            // to perform a dummy hash, making the execution time identical
-            // whether the user exists or not.
             let _ = security::hash_password(plaintext_password);
         }
 
-        // Generic error
         Err(LumeError::AccessDenied)
     }
 
-    /// Indexes a newly stored email message into the SQLite database.
-    ///
-    /// This records the metadata necessary for future retrieval, including the
-    /// `dict_id` required for decompression and the `acl_id` representing ownership.
     pub fn index_message(
         &self,
         message_id: &str,
@@ -116,15 +89,11 @@ impl LumeEngine {
         let db = self.db.lock().unwrap();
         db.execute(
             "INSERT INTO messages (message_id, acl_id, dict_id, subject, sender) VALUES (?1, ?2, ?3, ?4, ?5)",
-            params![message_id, header.acl_id, header.dict_id, subject, sender],
+            params![message_id, header.acl_id as i64, header.dict_id, subject, sender],
         )?;
         Ok(())
     }
 
-    /// Authorizes access to an email and retrieves the dictionary ID needed for decompression.
-    ///
-    /// Verifies that the requested `message_id` is owned by the provided `requesting_acl_id`.
-    /// Returns `LumeError::AccessDenied` on failure.
     pub fn authorize_and_get_dict(
         &self,
         message_id: &str,
@@ -133,7 +102,7 @@ impl LumeEngine {
         let db = self.db.lock().unwrap();
         let mut stmt =
             db.prepare("SELECT dict_id FROM messages WHERE message_id = ?1 AND acl_id = ?2")?;
-        let mut rows = stmt.query(params![message_id, requesting_acl_id])?;
+        let mut rows = stmt.query(params![message_id, requesting_acl_id as i64])?;
 
         if let Some(row) = rows.next()? {
             Ok(row.get(0)?)
@@ -142,7 +111,6 @@ impl LumeEngine {
         }
     }
 
-    /// Retrieves a list of message IDs matching a specific sender for a given user.
     pub fn search_by_sender(
         &self,
         requesting_acl_id: u64,
@@ -151,7 +119,7 @@ impl LumeEngine {
         let db = self.db.lock().unwrap();
         let mut stmt =
             db.prepare("SELECT message_id FROM messages WHERE acl_id = ?1 AND sender = ?2")?;
-        let mut rows = stmt.query(params![requesting_acl_id, sender])?;
+        let mut rows = stmt.query(params![requesting_acl_id as i64, sender])?;
 
         let mut results = Vec::new();
         while let Some(row) = rows.next()? {
@@ -160,7 +128,6 @@ impl LumeEngine {
         Ok(results)
     }
 
-    /// Retrieves a list of message IDs matching a specific subject (using LIKE) for a given user.
     pub fn search_by_subject(
         &self,
         requesting_acl_id: u64,
@@ -169,7 +136,10 @@ impl LumeEngine {
         let db = self.db.lock().unwrap();
         let mut stmt =
             db.prepare("SELECT message_id FROM messages WHERE acl_id = ?1 AND subject LIKE ?2")?;
-        let mut rows = stmt.query(params![requesting_acl_id, format!("%{}%", subject_query)])?;
+        let mut rows = stmt.query(params![
+            requesting_acl_id as i64,
+            format!("%{}%", subject_query)
+        ])?;
 
         let mut results = Vec::new();
         while let Some(row) = rows.next()? {
